+++ app/src/test/java/com/harrysoft/burstcoinexplorer/test/FileSizeUtilsTest.java
@@ -1,4 +1,4 @@
-package com.harrysoft.burstcoinexplorer;
+package com.harrysoft.burstcoinexplorer.test;
 
 import com.harrysoft.burstcoinexplorer.util.FileSizeUtils;
